Refactor test set up
- Split tests into multiple headers files to be included into the main
  test runner. Group of tests can be excluded by commenting out the
  include lines.

- Better logging of tests: include the file name and line number where
  the test case was defined.
